+++ firstlib/infra/fastapi/books.py
